worker/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
faster-whisper==1.0.3
torch==2.3.1
numpy>=1.24.0
soundfile>=0.12.1
pyannote.audio==3.3.1
